ballot contract

+++ AccountsDemo.sol
@@ -1,13 +1,13 @@
 pragma solidity ^0.8.11;
 
 contract AccountsDemo {
-  address public whoDeposited;
-  uint public depositAmt;
-  uint public accountBalance;
+    address public whoDeposited;
+    uint256 public depositAmt;
+    uint256 public accountBalance;
 
-  function deposit() public payable {
-    whoDeposited = msg.sender;
-    depositAmt = msg.value;
-    accountBalance = address(this).balance;
-  }
+    function deposit() public payable {
+        whoDeposited = msg.sender;
+        depositAmt = msg.value;
+        accountBalance = address(this).balance;
+    }
 }
